requirements.txt
flask
firebase_admin
gunicorn
python-dotenv
icecream
stripe
flask-cors
